refactor: 테스트의 static 메서드 추출
